+++ src/main/kotlin/racingcar/Main.kt
@@ -1,5 +1,7 @@
 package racingcar
 
+import racingcar.domain.RacingCarGame
+
 fun main() {
     RacingCarGame.start()
 }
