+++ src/main/kotlin/racingcar/domain/CarFactory.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 object CarFactory {
     fun create(count: Int): List<Car> {
