+++ src/main/kotlin/racingcar/ui/MessagePrinter.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.ui
 
 object MessagePrinter {
     fun print(message: String) {
