+++ src/test/kotlin/racingcar/domain/CarFactoryTest.kt
@@ -1,6 +1,5 @@
-package racingcar
+package racingcar.domain
 
-import org.assertj.core.api.Assertions
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.Test
