+++ src/test/kotlin/racingcar/domain/CarTest.kt
@@ -1,7 +1,6 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.*
-import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
