+++ src/test/kotlin/racingcar/domain/ResultGeneratorTest.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.Test
